test(unit): Reorganize the unit test modules
- Relocated all thumbnail utilities' tests to a new test module `thumb-utils.spec.mjs`
- Updated the test suites of thumbnail utilities to follow the namespace changes (30fbb44)
